sqlinjection_analyzer/requirements/production.txt
fastapi [standard]==0.112.2
clickhouse-connect==0.7.19
pydantic-settings==2.4.0
sqlparse==0.5.1
py_find_injection==0.1.1